rename main layer lib component
