--- libraries/mapping/layers/src/lib/layer-lib.tsx
@@ -1,5 +1,0 @@
-export {
-  default as LayerLib,
-  default,
-  type LibModalProps,
-} from "../components/NewLibModal";
